Rename main.c->queuetest.c
